backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
torch==2.1.0
transformers==4.35.0
numpy==1.24.3
requests==2.31.0
aiohttp==3.9.0
scikit-learn==1.3.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
python-dotenv==1.0.0
Django>=5.0
psycopg2-binary
django-unfold
httpx
# Optional: For advanced NLP
# spacy==3.7.0
# For search APIs (optional)
# google-search-results==2.4.2